Add files via upload
Trading with Momentum Project
In this project, I implemented a trading strategy on my own, and tested to see if it has the potential to be profitable. The project is supplied with a universe of stocks and time range and also with a textual description of how to generate a trading signal based on a momentum indicator. I computed the signal for the time range given and applied it to the dataset to produce projected returns. Finally, I performed a statistical test on the mean of the returns to conclude if there is alpha in the signal. For the dataset, I used the end of day from Quotemedia.